requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.1.13
django-cors-headers==4.7.0
django-filter==25.1
django-graphql-jwt==0.4.0
django-jazzmin==3.0.1
django-seed==0.3.1
Faker==37.3.0
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
mysql-connector-python==9.3.0
mysqlclient==2.2.7
pillow==11.2.1
promise==2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
